+++ packages/azure-themes/src/azure/AzureStyleSettings.ts
@@ -10,21 +10,24 @@ import { CommandBarStyles } from './styles/CommandBar.styles';
 import { CommandBarButtonStyles } from './styles/CommandBarButton.styles';
 import { CompoundButtonStyles } from './styles/CompoundButton.styles';
 import { ContextualMenuStyles } from './styles/ContextualMenu.styles';
+import { DatePickerStyles } from './styles/DatePicker.styles';
 import { DefaultButtonStyles } from './styles/DefaultButton.styles';
 import { DialogContentStyles, DialogFooterStyles } from './styles/Dialog.styles';
 import { DropdownStyles } from './styles/DropDown.styles';
 import { IconButtonStyles } from './styles/IconButton.styles';
 import { LabelStyles } from './styles/Label.styles';
 import { LinkStyles } from './styles/Link.styles';
+import { MessageBarStyles } from './styles/MessageBar.styles';
 import { ModalStyles } from './styles/Modal.styles';
 import { PanelStyles } from './styles/Panel.styles';
 import { PivotStyles } from './styles/Pivot.styles';
 import { PrimaryButtonStyles } from './styles/PrimaryButton.styles';
 import { ProgressIndicatorStyles } from './styles/ProgressIndicator.styles';
-import { TextFieldStyles } from './styles/TextField.styles';
 import { SearchBoxStyles } from './styles/SearchBox.styles';
 import { SliderStyles } from './styles/Slider.styles';
 import { SpinButtonStyles } from './styles/SpinButton.styles';
+import { TextFieldStyles } from './styles/TextField.styles';
+import { ToggleStyles } from './styles/Toggle.styles';
 
 // TODO: "any" is used here to get around "is using xxx but cannot be named" TS error. Should be able to remove
 //        this 'any' once we upgrade to TS3.1+
@@ -76,9 +79,9 @@ export const AzureStyleSettings = (theme: ITheme): any => {
     ContextualMenu: {
       styles: ContextualMenuStyles
     },
-    // DatePicker: {
-    //   styles: DatePickerStyles
-    // },
+    DatePicker: {
+      styles: DatePickerStyles
+    },
     DefaultButton: {
       styles: DefaultButtonStyles(theme)
     },
@@ -108,6 +111,9 @@ export const AzureStyleSettings = (theme: ITheme): any => {
     },
     Link: {
       styles: LinkStyles
+    },
+    MessageBar: {
+      styles: MessageBarStyles
     },
     Modal: {
       styles: ModalStyles
@@ -144,9 +150,9 @@ export const AzureStyleSettings = (theme: ITheme): any => {
     // },
     TextField: {
       styles: TextFieldStyles
+    },
+    Toggle: {
+      styles: ToggleStyles
     }
-    // Toggle: {
-    //   styles: ToggleStyles
-    // }
   };
 };