+++ packages/experiments/src/components/Accordion/Accordion.tsx
@@ -1,14 +1,15 @@
+/** @jsx withSlots */
 import * as React from 'react';
-import { createStatelessComponent, IStyleableComponentProps } from '../../Foundation';
-import { CollapsibleSection, ICollapsibleSectionProps, ICollapsibleSectionStyles } from '../../CollapsibleSection';
-import { IAccordionComponent, IAccordionProps, IAccordionStyles } from './Accordion.types';
+import { withSlots, getSlots } from '../../Foundation';
+import { createStatelessComponent } from '../../Foundation';
+import { CollapsibleSection, ICollapsibleSectionProps } from '../../CollapsibleSection';
+import { IAccordionComponent, IAccordionProps, IAccordionSlots, IAccordionStyles, IAccordionTokens } from './Accordion.types';
 import { styles } from './Accordion.styles';
 
-const AccordionItemType = (<CollapsibleSection /> as React.ReactElement<ICollapsibleSectionProps> &
-  IStyleableComponentProps<ICollapsibleSectionProps, ICollapsibleSectionStyles>).type;
+const AccordionItemType = (<CollapsibleSection /> as React.ReactElement<ICollapsibleSectionProps>).type;
 
 const view: IAccordionComponent['view'] = props => {
-  const { renderAs: RootType = 'div', classNames, collapseItems } = props;
+  const { collapseItems } = props;
 
   const children: React.ReactChild[] = React.Children.map(
     props.children,
@@ -28,7 +29,11 @@ const view: IAccordionComponent['view'] = props => {
     }
   );
 
-  return <RootType className={classNames.root}> {children} </RootType>;
+  const Slots = getSlots<ICollapsibleSectionProps, IAccordionSlots>(props, {
+    root: 'div'
+  });
+
+  return <Slots.root>{children}</Slots.root>;
 };
 
 const AccordionStatics = {
@@ -39,7 +44,7 @@ type IAccordionStatics = typeof AccordionStatics;
 
 export const Accordion: React.StatelessComponent<IAccordionProps> & {
   Item: React.StatelessComponent<ICollapsibleSectionProps>;
-} = createStatelessComponent<IAccordionProps, IAccordionStyles, IAccordionStatics>({
+} = createStatelessComponent<IAccordionProps, IAccordionStyles, IAccordionTokens, IAccordionStatics>({
   displayName: 'Accordion',
   styles,
   view,