+++ packages/theme-samples/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @uifabric/theme-samples
 
-This log was last generated on Thu, 13 Dec 2018 13:37:01 GMT and should not be manually modified.
+This log was last generated on Tue, 15 Jan 2019 13:36:45 GMT and should not be manually modified.
+
+## 0.1.4
+Tue, 15 Jan 2019 13:36:45 GMT
+
+### Patches
+
+- Convert Button themes to use tokens.
 
 ## 0.1.3
 Thu, 13 Dec 2018 13:37:01 GMT